Move all of the friends page code into the FriendsContainer to match the pattern of other pages
